Added services for saving/loading data.
Currently can't save data because of the PrimaryKey on the FieldValues_... tables also being a foreign key. EntityFramework tries to update that but can't because it's store-generated:

"A dependent property in a ReferentialConstraint is mapped to a store-generated column. Column: 'FieldValueID'."

If I really stop that being store generated, then it seems to go mad trying to re-save records in all the other tables, and violating their constrants.